requirements.txt
numpy
pandas
pika
python-dotenv
requests
